migrate system settings
